Add tests for additional types
