+++ src/ModelInfo.php
@@ -2,8 +2,6 @@
 
 namespace Spatie\ModelInfo;
 
-use Doctrine\DBAL\Exception;
-use Doctrine\DBAL\Platforms\AbstractPlatform;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Collection;
 use ReflectionClass;
@@ -11,22 +9,11 @@ use Spatie\ModelInfo\Attributes\Attribute;
 use Spatie\ModelInfo\Attributes\AttributeFinder;
 use Spatie\ModelInfo\Relations\Relation;
 use Spatie\ModelInfo\Relations\RelationFinder;
+use Spatie\ModelInfo\Util\RegistersAdditionalTypeMappings;
 
 class ModelInfo
 {
-    protected static array $typeMappings = [
-        'bit' => 'string',
-        'enum' => 'string',
-        'geometry' => 'string',
-        'geomcollection' => 'string',
-        'linestring' => 'string',
-        'multilinestring' => 'string',
-        'multipoint' => 'string',
-        'multipolygon' => 'string',
-        'point' => 'string',
-        'polygon' => 'string',
-        'sysname' => 'string',
-    ];
+    use RegistersAdditionalTypeMappings;
 
     /**
      * @param  string|null  $directory
@@ -120,27 +107,5 @@ class ModelInfo
         return $this->relations->first(
             fn (Relation $relation) => $relation->name === $name
         );
-    }
-
-    /**
-     * Register the custom Doctrine type mappings available in laravel
-     *
-     * @param  \Doctrine\DBAL\Platforms\AbstractPlatform  $platform
-     * @return void
-     *
-     * @throws Exception
-     *
-     * @see \Illuminate\Database\Console\DatabaseInspectionCommand::registerTypeMappings
-     */
-    protected static function registerTypeMappings(AbstractPlatform $platform): void
-    {
-        foreach (static::$typeMappings as $type => $value) {
-            $platform->registerDoctrineTypeMapping($type, $value);
-        }
-    }
-
-    public static function addTypeMapping(string $dbType, string $doctrineType): void
-    {
-        static::$typeMappings[$dbType] = $doctrineType;
     }
 }
