+++ tests/AttributeTest.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Support\Collection;
 use Spatie\ModelInfo\Attributes\AttributeFinder;
+use Spatie\ModelInfo\Tests\TestSupport\Models\ExtendedTypesModel;
 use Spatie\ModelInfo\Tests\TestSupport\Models\TestModel;
 use function Spatie\Snapshots\assertMatchesSnapshot;
 
@@ -22,6 +23,15 @@ it('can get virtual attribute php types of a model', function () {
     $this->assertNotNull($titleUppercaseAttr);
     $this->assertEquals('string', $titleUppercaseAttr->phpType);
     $this->assertEquals(null, $titleWithoutReturnTypeAttr->phpType);
+});
+
+it('can get extended column types for a model', function () {
+    AttributeFinder::addTypeMapping('time', 'datetime');
+    $attributes = AttributeFinder::forModel(new ExtendedTypesModel());
+
+    expect($attributes)->toHaveCount(6);
+
+    matchesAttributesSnapshot($attributes);
 });
 
 function matchesAttributesSnapshot(Collection $attributes)
