+++ tests/TestSupport/Models/ExtendedTypesModel.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Spatie\ModelInfo\Tests\TestSupport\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class ExtendedTypesModel extends Model
+{
+}
